Refactor: Table
moved helper function to idGenerator.ts, interfaces to dataTypes.ts, and subcomponents of the table to TableSucomponents.tsx
